Sign In & Sign Up adding

+++ Library Managment Project/DTOs/MemberDTOs/AddMember.Request.cs
@@ -5,13 +5,13 @@ namespace Library_Managment_Project.DTOs.MemberDTOs
     public class AddMemberRequest
     {
         public int MemberId { get; set; }
-        public long MemberNumber { get; set; }
+        public int MemberNumber { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
         public int PhoneNumber {  get; set; }
         public  MemberShipType MemberShipType { get; set; }
-        public AddMemberRequest(int memberId, long memberNumber, string firstName, string lastName, string email, int phoneNumber , MemberShipType memberShipType)
+        public AddMemberRequest(int memberId, int memberNumber, string firstName, string lastName, string email, int phoneNumber , MemberShipType memberShipType)
         {
             MemberId = memberId;
             MemberNumber = memberNumber;
