+++ Library Managment Project/Entities/Member.cs
@@ -4,7 +4,7 @@ namespace Library_Managment_Project.Entities
 {
     public class Member : User
     {
-        public long MemberCode { get; set; }
+        public int MemberCode { get; set; }
         public MemberShipType MemberShipType { get; set; }
         public ICollection<LoansBook> Loans { get; set; }
     }
